refactor: domain, view 파일로 분리
