+++ __tests__/CountOfAttempt.test.js
@@ -1,4 +1,4 @@
-import { Validator } from '../src/models/Validator.js';
+import { Validator } from '../src/domain/models/Validator.js';
 
 describe('시도 횟수 입력에 대한 유효성 테스트', () => {
 	test('type이 number이고 자연수이며 시도 제한 횟수 이하의 유효한 값을 넣었을 때 해당 값을 반환한다.', () => {
