+++ __tests__/ScoreBoard.test.js
@@ -1,4 +1,4 @@
-import ScoreBoard from '../src/models/ScoreBoard.js';
+import ScoreBoard from '../src/domain/models/ScoreBoard.js';
 
 describe('스코어보드에 대한 유효성 테스트', () => {
 	test('중복된 이름이 없으며 각 이름이 5글자 이하 및 특수문자가 존재하지 않는 유효한 값을 넣으면, 0으로 초기화된 Map 객체를 반환한다', () => {
