+++ src/domain/models/ScoreBoard.js
